Move more constants out of Constants
