+++ core/src/main/java/com/microsoft/applicationinsights/internal/statsbeat/OperatingSystem.java
@@ -1,0 +1,17 @@
+package com.microsoft.applicationinsights.internal.statsbeat;
+
+enum OperatingSystem {
+
+    OS_WINDOWS("Windows"), OS_LINUX("Linux"), OS_UNKNOWN("unknown");
+
+    private final String id;
+
+    OperatingSystem(String id) {
+        this.id = id;
+    }
+
+    @Override
+    public String toString() {
+        return id;
+    }
+}
