+++ core/src/main/java/com/microsoft/applicationinsights/internal/statsbeat/ResourceProvider.java
@@ -1,0 +1,21 @@
+package com.microsoft.applicationinsights.internal.statsbeat;
+
+enum ResourceProvider {
+
+    RP_FUNCTIONS("functions"),
+    RP_APPSVC("appsvc"),
+    RP_VM("vm"),
+    RP_AKS("aks"),
+    UNKNOWN("unknown");
+
+    private final String id;
+
+    ResourceProvider(String id) {
+        this.id = id;
+    }
+
+    @Override
+    public String toString() {
+        return id;
+    }
+}
